requirements.txt
autopep8==1.5.6
backcall==0.2.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
easygui==0.98.2
et-xmlfile==1.0.1
imageio==2.9.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
openpyxl==3.0.6
pandas==1.2.2
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
pycodestyle==2.7.0
Pygments==2.8.0
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PySide2==5.15.2
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
pywin32==300
pyzmq==22.0.3
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
shiboken2==5.15.2
six==1.15.0
toml==0.10.2
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
xlwt==1.3.0
